requirements.txt
Flask==2.0.3
requests==2.27.1
mysql-connector-python==8.0.27
Werkzeug==2.0.3
pika==1.2.0